chore: auth, users api

+++ controllers/users.js
@@ -1,17 +1,28 @@
+const Post = require("../models/post");
 const User = require("../models/user");
 
-exports.getUser = async (req, res, next) => {
-  const { id } = req.params;
-  console.log("결과 :", Number(id) === req?.user?.id);
-
+exports.getMe = async (req, res, next) => {
   try {
-    if ((Number(id) === req?.user?.id) === false) {
-      return res.status(401).send({ error: "Unauthorized" });
+    if (req?.user) {
+      const userWithoutPwd = await User.findOne({
+        where: { id: req.user?.id },
+        attributes: {
+          exclude: ["password", "provider"],
+        },
+        include: [
+          {
+            model: Post,
+            attributes: ["id"],
+          },
+          { model: User, as: "Followings", attributes: ["id"] },
+          { model: User, as: "Followers", attributes: ["id"] },
+        ],
+      });
+      console.log("userWithoutPwd :", userWithoutPwd);
+      return res.status(200).json(userWithoutPwd);
+    } else {
+      return res.status(200).json(null);
     }
-    const user = await User.findByPk(id);
-    return res
-      .status(200)
-      .json({ userId: user?.id, email: user?.email, nickname: user?.nickname });
   } catch (err) {
     console.error(err);
     return next(err);
@@ -19,17 +30,17 @@ exports.getUser = async (req, res, next) => {
 };
 
 exports.patchUser = async (req, res, next) => {
-  const { id } = req.params;
   const body = req.body;
   try {
-    if ((Number(id) === req?.user?.id) === false) {
-      return res.status(401).send({ error: "Unauthorized" });
+    if (req?.user) {
+      await User.update(
+        { nickname: body.nickname },
+        { where: { id: req.user.id } }
+      );
+      return res.status(200).json({ nickname: body.nickname });
+    } else {
+      return res.status(200).json({ error: "Unauthorized" });
     }
-    await User.update(
-      { nickname: body.nickname },
-      { where: { id: req.user.id } }
-    );
-    return res.status(200).json({ nickname: body.nickname });
   } catch (err) {
     console.error(err);
     return next(err);
@@ -38,17 +49,53 @@ exports.patchUser = async (req, res, next) => {
 
 exports.deleteUser = async (req, res, next) => {
   try {
-    await User.destroy({
-      where: { id: req?.user?.id },
+    if (req?.user) {
+      await User.destroy({
+        where: { id: req?.user?.id },
+      });
+      res.clearCookie("connect.sid");
+      res.clearCookie("session-cookie");
+      await req.session.destroy((err) => {
+        if (err) {
+          return res.status(400).send("로그인하지 않았습니다.");
+        }
+      });
+      return res.status(200).send("회원탈퇴 처리 완료");
+    } else {
+      return res.status(200).json({ error: "Unauthorized" });
+    }
+  } catch (err) {
+    console.error(err);
+    return next(err);
+  }
+};
+
+/**  /users/1  */
+exports.getUser = async (req, res, next) => {
+  try {
+    const userWithoutPwd = await User.findOne({
+      where: { id: req?.params.id },
+      attributes: {
+        exclude: ["password"],
+      },
+      include: [
+        {
+          model: Post,
+          attributes: ["id"],
+        },
+        { model: User, as: "Followings", attributes: ["id"] },
+        { model: User, as: "Followers", attributes: ["id"] },
+      ],
     });
-    res.clearCookie("connect.sid");
-    res.clearCookie("session-cookie");
-    await req.session.destroy((err) => {
-      if (err) {
-        return res.status(400).send("로그인하지 않았습니다.");
-      }
-    });
-    return res.status(200).send("회원탈퇴 처리 완료");
+    if (userWithoutPwd) {
+      const userData = userWithoutPwd.toJSON();
+      userData.Posts = userData.Posts.length;
+      userData.Followers = userData.Followers.length;
+      userData.Followings = userData.Followings.length;
+      return res.status(200).json(userData);
+    } else {
+      return res.status(404).json("존재하지 않는 회원입니다.");
+    }
   } catch (err) {
     console.error(err);
     return next(err);
